==> project/0-download-data.ipynb <==
### How to download the data

1. Download the data from here: https://drive.google.com/file/d/1c4yLdc-rmmBwSAv4FHiFBPrO1io0SgLR/view?usp=sharing
2. Alternatively the data can be downloaded from here (but access needs to be requested): https://drive.google.com/drive/u/0/folders/0B_D5tuT1vDQtckFGWkk1aTh5VlE?resourcekey=0-pu1i9MdbQY1IzfOlsQJOQA
3. Unzip the file and place all files in the `data` folder.
