==> project/1-3-preprocessing_shared_interest_detection_data.ipynb <==
import pandas as pd

# Load services data

service_columns = ['User_Id', 'Service_Id', 'Service_Name', 'Service_URL', 'UserName_On_Service', 'UserURL_On_Service']

df_services = pd.read_csv('data/services.csv', delimiter='|', on_bad_lines='skip', names=service_columns, usecols=[0, 1, 2, 3, 4, 5], header=None)

df_services.shape

# Load valid followings data

community_detection_data = pd.read_pickle("community_detection_data.pkl")

df_followings = community_detection_data['followings']

valid_users_in_community = set(df_followings['Follower_Id']).union(df_followings['Followed_Id'])

len(valid_users_in_community)

# Clean up services data

df_services.dropna(subset=['User_Id'], inplace=True)
df_services.dropna(subset=['Service_Id'], inplace=True)
df_services.drop_duplicates(inplace=True)

df_services['User_Id'] = df_services['User_Id'].str.lower()

df_services['Service_Id'] = df_services['Service_Id'].str.lower()

df_services = df_services[
                df_services['User_Id'].isin(valid_users_in_community)
            ]

df_services.shape

# Load subscriptions data

subscription_columns = ['Subscriber_Id', 'Subscribed_To_Id']

df_subscriptions = pd.read_csv('data/subscriptions.csv', delimiter=',', names=subscription_columns)

data_services = { 'services': df_services }
pd.to_pickle(data_services, 'shared_interest_detection_data_services.pkl')
