==> project/1-4-preprocessing_user_activity_data.ipynb <==
import pandas as pd

# Load Likes data

likes_columns = ['User_Id', 'Post_Id']

df_likes = pd.read_csv('./data/likes.csv', delimiter='\t', usecols=[0, 1], names=likes_columns)

df_likes.shape

# Load posts data

post_columns = ['Post_Id', 'User_Id', 'Service_Name', 'Text', 'NumImg', 'NumVdo']

df_posts_1 = pd.read_csv('data/entries1.csv', delimiter='\t', on_bad_lines='skip', usecols=[0, 1, 2, 7, 8, 10], names=post_columns, header=None)

df_posts_1.shape

df_posts_2 = pd.read_csv('./data/entries2.csv', delimiter='\t', on_bad_lines='skip', usecols=[0, 1, 2, 7, 8, 10], names=post_columns, header=None)

df_posts_2.shape

df_posts_3 = pd.read_csv('./data/entries3.csv', delimiter='\t', on_bad_lines='skip', usecols=[0, 1, 2, 7, 8, 10], names=post_columns, header=None)

df_posts_3.shape

# Load comments data

comments_columns = ['Post_Id', 'Entry_Id', 'Posted_By', 'Source_Name']

df_comments = pd.read_csv('./data/comments.csv', on_bad_lines='skip', delimiter='\t', usecols=[0, 1, 2, 3], names=comments_columns,  header=None)

df_comments.shape

# Load valid followings data

community_detection_data = pd.read_pickle("community_detection_data.pkl")

df_followings = community_detection_data['followings']

valid_users_for_activities = set(df_followings['Follower_Id']).union(df_followings['Followed_Id'])

# Clean up posts data

df_posts_1.dropna(subset=['Text', 'NumImg', 'NumVdo'], how='all', inplace=True)

df_posts_1.shape

df_posts_2.dropna(subset=['Text', 'NumImg', 'NumVdo'], how='all', inplace=True)

df_posts_2.shape

df_posts_3.dropna(subset=['Text', 'NumImg', 'NumVdo'], how='all', inplace=True)

df_posts_3.shape

df_posts_1['User_Id'] = df_posts_1['User_Id'].str.lower()

df_posts_1['Post_Id'] = df_posts_1['Post_Id'].str.lower()

df_posts_2['User_Id'] = df_posts_2['User_Id'].str.lower()

df_posts_2['Post_Id'] = df_posts_2['Post_Id'].str.lower()

df_posts_3['User_Id'] = df_posts_3['User_Id'].str.lower()

df_posts_3['Post_Id'] = df_posts_3['Post_Id'].str.lower()

df_valid_posts_1 = df_posts_1[
    df_posts_1['User_Id'].isin(valid_users_for_activities)
]

df_valid_posts_1.head()

df_valid_posts_2 = df_posts_2[
    df_posts_2['User_Id'].isin(valid_users_for_activities) 
]

df_valid_posts_2.head()

df_valid_posts_3 = df_posts_3[
    df_posts_3['User_Id'].isin(valid_users_for_activities) 
]

df_valid_posts_3.head()

df_valid_posts = pd.concat([df_valid_posts_1[['Post_Id', 'User_Id', 'Service_Name']],
                            df_valid_posts_2[['Post_Id', 'User_Id', 'Service_Name']],
                            df_valid_posts_3[['Post_Id', 'User_Id', 'Service_Name']]])

df_valid_posts.shape

df_valid_posts.head()

# Clean up likes data

df_likes.drop_duplicates(inplace=True)

df_likes['User_Id'] = df_likes['User_Id'].str.lower()
df_likes['Post_Id'] = df_likes['Post_Id'].str.lower()

df_likes.shape

df_valid_likes = df_likes[df_likes['User_Id'].isin(valid_users_for_activities)]

df_valid_likes.head()

# Clean up comments data

df_comments.dropna(subset=['Posted_By'], inplace=True)

df_comments.drop_duplicates(inplace=True)

df_comments['Posted_By'] = df_comments['Posted_By'].str.lower()

df_comments.shape

df_valid_comments = df_comments[df_comments['Posted_By'].isin(valid_users_for_activities)]

df_comments.head()

# Export data

likes_data = { 'likes': df_valid_likes }

pd.to_pickle(likes_data, 'user_activity_data_likes.pkl')

comments_data = { 'comments': df_valid_comments }

pd.to_pickle(comments_data, 'user_activity_data_comments.pkl')

posts_data = { 'posts': df_valid_posts }

pd.to_pickle(posts_data, 'user_activity_data_posts.pkl')
