==> project/2-1-social-circle-detection.ipynb <==
import pandas as pd

data_followings = pd.read_pickle("community_detection_data.pkl")
df_followings = data_followings['followings']

# Leiden community detection

import igraph as ig
import leidenalg as la

edges = list(zip(df_followings['Follower_Id'], df_followings['Followed_Id']))

grph = ig.Graph.TupleList(edges, directed=True)

objective_function = la.RBConfigurationVertexPartition

# Best possible solution based on resolution

resolutions = [0.5, 1, 1.5, 2, 2.5, 3]

for resl in resolutions:

    test_partition_leiden = la.find_partition(
        grph, 
        objective_function, 
        resolution_parameter=resl,
        n_iterations=10,
        seed=42
    )

    test_modularity_leiden = grph.modularity(test_partition_leiden)

    print(f"For resolution: {resl}, Leiden found {len(test_partition_leiden)} communities with modularity {test_modularity_leiden:.4f}")

resolution = 1.0 

# Execute Leiden

partition = la.find_partition(
    grph, 
    objective_function, 
    resolution_parameter=resolution,
    n_iterations=10,
    seed=42
)

modularity = grph.modularity(partition)

print(f"Leiden found {len(partition)} communities with modularity {modularity:.4f}")

df_communities = []
adj_outgoing_edges_leiden = {v.index: set(grph.successors(v.index)) for v in grph.vs}

for i, community_nodes in enumerate(partition):
    subgraph = grph.subgraph(community_nodes)

    nodes = subgraph.vs.indices
    nodes_set = set(nodes)
    
    internal_density = subgraph.density()
    
    edges_in_len = len(subgraph.es)
    edges_out_len = sum(len(adj_outgoing_edges_leiden[node] - nodes_set) for node in nodes)
    
    if (edges_in_len + edges_out_len) > 0:
        conductance = edges_out_len / (edges_in_len + edges_out_len)
    else:
        conductance = 0
        
    df_communities.append({
        'Community': i,
        'Size': len(community_nodes),
        'Density': internal_density,
        'Conductance': conductance 
    })

df_community_summary = pd.DataFrame(df_communities).sort_values(by='Size', ascending=False)

df_community_summary

df_community_summary.describe()

# Users in communities

membership_leiden = partition.membership

df_community_users = pd.DataFrame({
    "Community": membership_leiden,
    "User_Id": grph.vs['name']
})

df_community_users

df_communities = df_community_users.groupby('Community')['User_Id'].size().reset_index(name='Size')

df_communities

df_communities.sort_values('Size', ascending=False)

community_size_min_threshold = 4 # As 75% communities are <= 3
df_communities_filtered = df_communities[df_communities['Size'] >= community_size_min_threshold]

df_communities_filtered


# Export community data

community_data = { 'Communities': df_communities, 'Communities_filtered': df_communities_filtered }

pd.to_pickle(community_data, 'community_data.pkl')

community_user_data = { 'Users': df_community_users }

pd.to_pickle(community_user_data, 'community_users_data.pkl')
