==> project/2-2-community_detection_models.ipynb <==
import pandas as pd

data = pd.read_pickle("community_detection_data.pkl")

df_followings = data['followings']

df_followings.head()

# Community detection model

import igraph as ig

edges = list(zip(df_followings['Follower_Id'], df_followings['Followed_Id']))

Model 1: Leiden Method

grph_leiden = ig.Graph.TupleList(edges, directed=True)

grph_leiden.simplify() 

print(f"Number of nodes (users): {grph_leiden.vcount()}")
print(f"Number of edges (connections): {grph_leiden.ecount()}")

import leidenalg as la

partition_leiden = la.find_partition(
    grph_leiden, 
    la.RBConfigurationVertexPartition, 
    resolution_parameter=1,
    n_iterations=10,
    seed=42
)

modularity_leiden = grph_leiden.modularity(partition_leiden)

print(f"Leiden found {len(partition_leiden)} communities with modularity {modularity_leiden:.4f}")

Model 2: Louvian Method

grph_louvain = ig.Graph.TupleList(edges, directed=False)

grph_louvain.simplify()

partition_louvain = grph_louvain.community_multilevel()

modularity_louvain = grph_louvain.modularity(partition_louvain)

print(f"Louvain found {len(partition_louvain)} communities with modularity {modularity_louvain:.4f}")

# Leiden model evaluation

Modularity

modularity_leiden = grph_leiden.modularity(partition_leiden)

print(f"Leiden found {len(partition_leiden)} communities with modularity {modularity_leiden:.4f}")

adj_outgoing_edges_leiden = {v.index: set(grph_leiden.successors(v.index)) for v in grph_leiden.vs}    

leiden_community_metrics = []


for community, subgraph in enumerate(partition_leiden.subgraphs()):
    density = subgraph.density()
    avg_deg = sum(subgraph.degree()) / len(subgraph.vs)

    nodes = subgraph.vs.indices
    nodes_set = set(nodes)
    
    internal_edges = sum(len(adj_outgoing_edges_leiden[node] & nodes_set) for node in nodes)
    cut_edges = sum(len(adj_outgoing_edges_leiden[node] - nodes_set) for node in nodes)
    
    total_edges = internal_edges + cut_edges
    conductance = cut_edges / total_edges if total_edges > 0 else 0

    leiden_community_metrics.append({
        "community": community,
        "size": len(nodes),
        "density": density,
        "avg_degree": avg_deg,
        "internal_edges": internal_edges,
        "cut_edges": cut_edges,
        "conductance": conductance
    })

df_leiden_community_metrics = pd.DataFrame(leiden_community_metrics, columns=['community', 'size', 'density', 'avg_degree', 'internal_edges', 'cut_edges', 'conductance'])

df_leiden_community_metrics.describe()

# Louvain Model Evaluation

modularity_louvain = grph_louvain.modularity(partition_louvain)

print(f"Louvain found {len(partition_louvain)} communities with modularity {modularity_louvain:.4f}")

adj_edges_louvian = {v.index: set(grph_louvain.neighbors(v)) for v in grph_louvain.vs}

louvain_community_metrics = []

for community, subgraph in enumerate(partition_louvain.subgraphs()):
    density = subgraph.density()
    avg_deg = sum(subgraph.degree()) / len(subgraph.vs)
    
    nodes = subgraph.vs.indices
    nodes_set = set(nodes)
    
    # Internal edges: edges fully inside the community
    internal_edges = sum(len(adj_edges_louvian[node] & nodes_set) for node in nodes) / 2
    
    # Cut edges: edges leaving the community
    cut_edges = sum(len(adj_edges_louvian[node] - nodes_set) for node in nodes) / 2
    
    total_edges = internal_edges + cut_edges
    conductance = cut_edges / total_edges if total_edges > 0 else 0
    
    louvain_community_metrics.append({
        "community": community,
        "size": len(nodes),
        "density": density,
        "avg_degree": avg_deg,
        "internal_edges": internal_edges,
        "cut_edges": cut_edges,
        "conductance": conductance
    })

df_louvain_community_metrics = pd.DataFrame(louvain_community_metrics, columns=['community', 'size', 'density', 'avg_degree', 'internal_edges', 'cut_edges', 'conductance'])

df_louvain_community_metrics.describe()
