==> project/2-3-characterize_communities_by_services.ipynb <==
import pandas as pd
import numpy as np

# Load community data

community_data = pd.read_pickle('community_data.pkl')

df_communities = community_data['Communities']

df_communities.shape

df_communities_filtered = community_data['Communities_filtered']

df_communities_filtered.shape

community_user_data = pd.read_pickle('community_users_data.pkl')

df_community_users = community_user_data['Users']

df_community_users.shape

# Load services data

services_data = pd.read_pickle('shared_interest_detection_data_services.pkl')

df_services = services_data['services']

df_services.head()

df_community_user_services = df_community_users.merge(df_services[['User_Id', 'Service_Id']], how='left', on='User_Id')

df_community_user_services.head()

df_community_services = df_community_user_services.groupby(['Community', 'Service_Id']).count().reset_index().rename({ 'User_Id': 'Service_Count' }, axis=1).sort_values(['Community', 'Service_Count'], ascending=[True, False]).rename({ 'Service_Id': 'Service' }, axis = 1)

df_community_services.describe()

df_community_services

df_community_service_usage = df_community_services.groupby('Community')['Service'].count().reset_index().rename({ 'Service': 'Service_Count' }, axis=1)

df_community_service_usage.sort_values('Service_Count', ascending=[False])

df_community_service_usage = df_community_service_usage.merge(df_communities, on='Community').rename({ 'Size': 'User_Count' }, axis=1)

df_community_service_usage.describe()

df_communities

df_community_top_services = df_community_services.groupby(['Community']).head(3)

df_community_top_services

df_community_top_services_agg = df_community_top_services.groupby('Community').apply(lambda x: set(x.Service)).reset_index(name='Services')

df_community_top_services_agg

# Cluster similar communities based on top services

filtered_community_ids = df_communities_filtered['Community']

filtered_community_ids

df_community_top_services_filtered = df_community_top_services[df_community_top_services['Community'].isin(filtered_community_ids)]

df_community_top_services_filtered

df_community_top_services_filtered_pivot = pd.pivot_table(
    df_community_top_services_filtered,
    index='Community',
    columns='Service',
    aggfunc='count',
    fill_value=0
)

df_community_top_services_filtered_pivot

Jaccard Similarity Matrix

from scipy.spatial.distance import pdist, squareform

jaccard_distances_vector = pdist(df_community_top_services_filtered_pivot.values, metric='jaccard')

jaccard_distances_vector

print(np.min(jaccard_distances_vector))
print(np.max(jaccard_distances_vector))

condensed_dist_matrix = squareform(jaccard_distances_vector)

from scipy.cluster.hierarchy import linkage
import matplotlib.pyplot as plt
from sklearn.cluster import AgglomerativeClustering
from sklearn.metrics import silhouette_score
from scipy.cluster.hierarchy import fcluster

Analyze distance threshold

Z = linkage(jaccard_distances_vector, method='complete')

merge_distances = Z[:, 2]

candidate_thresholds = np.unique(merge_distances)

candidate_thresholds

for dt in candidate_thresholds:

    labels = fcluster(Z, dt, criterion='distance')
    
    num_clusters = len(np.unique(labels))
    
    if num_clusters >= 2:
        score = silhouette_score(condensed_dist_matrix, labels, metric='precomputed')

        print(f"Silhouette score: {score:.4f} for distance threshold: {dt} and clusters count: {num_clusters}")
    


Apply agglomerative clustering

chosen_k = 94 # considering K from slightly reduced silhoutte score for meaningful clusters

clustering_model = AgglomerativeClustering(
    n_clusters=chosen_k,
    linkage='complete',
    metric='precomputed'
)
    
cluster_labels = clustering_model.fit_predict(condensed_dist_matrix)

score = silhouette_score(
    condensed_dist_matrix,
    cluster_labels,
    metric='precomputed'
)

print(f"Optimal number of clusters {chosen_k} suggested by Silhouette Score: {score}")

df_community_top_services_filtered_agg = df_community_top_services_filtered.groupby('Community').apply(lambda x: set(x.Service)).reset_index(name='Services')

df_community_top_services_filtered_agg

df_community_top_services_filtered_agg['Cluster'] = cluster_labels

df_community_top_services_filtered_agg

df_community_cluster_features = df_community_top_services_filtered_agg.explode('Services').groupby(['Cluster', 'Services']).mean().reset_index().rename({ 'Community': 'Mean' }, axis=1).sort_values(['Cluster', 'Mean'], ascending=[True, False])

df_community_cluster_features

top_feature_count = 3
cluster_summaries = []

for cluster_id in range(chosen_k):

    top_features = df_community_cluster_features[df_community_cluster_features['Cluster'] == cluster_id]['Services'].head(top_feature_count)

    cluster_summaries.append({
        'Cluster': cluster_id,
        'Top Services': set(top_features),
        'Communitiy_Count': len(df_community_top_services_filtered_agg[df_community_top_services_filtered_agg['Cluster'] == cluster_id])
    })


cluster_profiles = pd.DataFrame(cluster_summaries)

cluster_profiles.sort_values('Communitiy_Count', ascending=False).head(5)

cluster_profiles.sort_values('Communitiy_Count', ascending=True).head(5)

len(set(df_community_cluster_features['Services'].tolist()))

service_interpretation = {
    # Mainstream Social Media
    'facebook': 'Mainstream Social Media',
    'twitter': 'Mainstream Social Media',
    'youtube': 'Mainstream Social Media',
    
    # Blogging / Content Sharing
    'blog': 'Blogging ',
    'tumblr': 'Blogging ',
    'livejournal': 'Blogging',
    'plurk': 'Blogging',
    'backtype': 'Blogging',
    
    # Social Bookmarking / Curation
    'delicious': 'Social Bookmarking ',
    'digg': 'Social Bookmarking ',
    'diigo': 'Social Bookmarking',
    'misterwong': 'Social Bookmarking',
    'stumbleupon': 'Social Bookmarking ',
    
    # Professional / Networking
    'linkedin': 'Professional',
    'googletalk': 'Professional ',
    'identica': 'Professional',
    
    # Entertainment / Media
    'lastfm': 'Entertainment',
    'netflix': 'Entertainment ',
    'pandora': 'Entertainment ',
    'mixx': 'Entertainment',
    
    # Photography / Visual Sharing
    'flickr': 'Photography ',
    'picasa': 'Photography',
    'photobucket': 'Photography ',
    
    # Feed / Reader Platforms
    'feed': 'Feed ',
    'googlereader': 'Feed ',
    'googleshared': 'Feed',
    
    # Other / Niche
    'amazon': 'Other',
    'ameba': 'Other',
    'brightkite': 'Other',
    'disqus': 'Other',
}

def get_cluster_interpretations(top_services):
    return list({service_interpretation.get(s, 'Other / Misc') for s in top_services})

cluster_profiles['Interpretation'] = cluster_profiles['Top Services'].apply(get_cluster_interpretations)
cluster_profiles

cluster_profiles.sort_values('Communitiy_Count', ascending=True)

cluster_profiles.describe()
